% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alto.R
\name{coords_xy}
\alias{coords_xy}
\title{Parse coords of points}
\usage{
coords_xy(data, split = ",")
}
\arguments{
\item{data}{a character vector or a list of character data}

\item{split}{which split character to use between x and y. Defaults to ','}
}
\value{
a list of data.frames with columns x and y
}
\description{
Parse coords of points where the list of coordinates is available in a string
as follows: x1,y1 x2,y2, x3,y3
}
\examples{
x <- c("0.1,0.2 0.3,0.15", "", NA, "1,2 3,4 5,6 1,0")
coords_xy(x)
x <- c("0.1,0.2 0.3,0.15", "", NA, "1,2 3,4 5,6 1,0")
x <- strsplit(x, " ")
coords_xy(x)
}
\keyword{internal}
