% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_crop_area}
\alias{image_crop_area}
\title{Extract a polygonal region}
\usage{
image_crop_area(image, x, bbox = FALSE, ...)
}
\arguments{
\item{image}{\code{image} either an object of class \code{opencv-image} or a path to an image file on disk}

\item{x}{a list vector where each list element contains columns x and y indicating the positions of the points of the polygon
or a single data.frame with columns \code{x} and \code{y}}

\item{bbox}{logical indicating to limit the result to the bounding box of the polygon}

\item{...}{further arguments currently not used}
}
\value{
a \code{magick-image} or a list of \code{magick-image} objects
}
\description{
Extract a polygonal region, optionally limiting the extracting to the bounding box
}
\examples{
library(opencv)
path     <- system.file(package = "madoc.utils", "extdata", "pagexml-example.xml")
x        <- read_pagexml(path)
x
img      <- system.file(package = "madoc.utils", "extdata", "pagexml-example.jpg")
img      <- ocv_read(img)
pts      <- x$points[[1]]
area     <- image_crop_area(img, pts, bbox = FALSE)
area
area     <- image_crop_area(img, pts, bbox = TRUE)
area
areas    <- image_crop_area(img, x$points, bbox = FALSE)
}
