% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_crop_baselineareas}
\alias{image_crop_baselineareas}
\title{Extract areas between baselines}
\usage{
image_crop_baselineareas(
  image,
  x,
  extend = TRUE,
  color = "royalblue",
  border = "10x10",
  overview = TRUE,
  max_width = +Inf,
  trace = FALSE,
  ...
)
}
\arguments{
\item{image}{\code{image} either an object of class \code{opencv-image} or a path to an image file on disk}

\item{x}{a list vector where each list element contains columns x and y indicating the positions of the baseline}

\item{extend}{logical indicating to extend the baseline to the left and right of the image. Defaults to TRUE.}

\item{color}{color to use for adding a border in the overview image. Defaults to 'royalblue'.}

\item{border}{border pixels to using in the overview image. Defaults to 10x10 pixel borders.}

\item{overview}{logical indicating to add the overview image of all area's below each other. Defaults to TRUE.}

\item{max_width}{maximum width of the overview image. Defaults to +Inf}

\item{trace}{logical indicating to trace progress}

\item{...}{further arguments currently not used}
}
\value{
a list with elements areas and overview where \code{overview} is a \code{magick-image} with stacked image lines
and \code{areas} is a list of \code{magick-image}'s, one for each text line \cr
In case overview is set to \code{FALSE} the return value is only the list of stacked image lines.
}
\description{
Extract areas between baselines
}
\examples{
library(opencv)
library(magick)
path     <- system.file(package = "madoc.utils", "extdata", "pagexml-example.xml")
x        <- read_pagexml(path)
x
img      <- system.file(package = "madoc.utils", "extdata", "pagexml-example.jpg")
img      <- ocv_read(img)
areas    <- image_crop_baselineareas(img, x = x$baseline, extend = FALSE, trace = TRUE)
areas$areas
image_resize(areas$overview, "x600")
areas    <- image_crop_baselineareas(img, x = x$baseline, extend = TRUE, color = "red")
image_resize(areas$overview, "x600")
}
