% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_crop_textlines}
\alias{image_crop_textlines}
\title{Crop out areas with text lines from an image}
\usage{
image_crop_textlines(image, geometries, color = "royalblue", border = "10x10")
}
\arguments{
\item{image}{\code{image} either an object of class \code{magick-image} or a path to an image file on disk}

\item{geometries}{a data.frame with columns width, height, x_left, y_top indicating the areas to extract from the image}

\item{color}{color to use for adding a border in the overview image. Defaults to 'royalblue'.}

\item{border}{border pixels to using in the overview image. Defaults to 10x10 pixel borders.}
}
\value{
a list with elements areas and overview where \code{overview} is a \code{magick-image} with stacked image lines
and \code{areas} is a list of \code{magick-image}'s, one for each text line
}
\description{
Crop out areas with text lines from an image
}
\examples{
library(magick) 
library(data.table)
path     <- system.file(package = "madoc.utils", "extdata", "alto-example.xml")
x        <- read_alto(path)
x$width  <- as.integer(x$WIDTH)
x$height <- as.integer(x$HEIGHT)
x$x_left <- as.integer(x$HPOS)
x$y_top  <- as.integer(x$VPOS)

img      <- system.file(package = "madoc.utils", "extdata", "alto-example.jpg")
img      <- image_read(img)
areas    <- image_crop_textlines(img, x, color = "red")
areas$overview
areas$areas
}
