% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_draw_baselines}
\alias{image_draw_baselines}
\title{Draw baselines on an image}
\usage{
image_draw_baselines(image, x, ...)
}
\arguments{
\item{image}{\code{image} either an object of class \code{magick-image} or a path to an image file on disk}

\item{x}{a list vector where each list element contains columns x and y indicating the positions of the baseline}

\item{...}{further arguments passed on to \code{\link{lines}}}
}
\value{
a \code{magick-image}
}
\description{
Draw baselines on an image
}
\examples{
library(magick)
path     <- system.file(package = "madoc.utils", "extdata", "pagexml-example.xml")
x        <- read_pagexml(path)
x

img      <- system.file(package = "madoc.utils", "extdata", "pagexml-example.jpg")
img      <- image_read(img)
plt      <- image_draw_baselines(img, x$baseline)
plt
plt      <- image_draw_baselines(img, x$baseline, col = "red", lwd = 10, lty = 2)
plt
}
