% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_draw_polygons}
\alias{image_draw_polygons}
\title{Draw polygons on an image}
\usage{
image_draw_polygons(image, x, ...)
}
\arguments{
\item{image}{\code{image} either an object of class \code{magick-image} or a path to an image file on disk}

\item{x}{a list vector where each list element contains columns x and y indicating the positions of the polygon}

\item{...}{further arguments passed on to \code{\link{polygon}}}
}
\value{
a \code{magick-image}
}
\description{
Draw polygons on an image
}
\examples{
library(magick)
path     <- system.file(package = "madoc.utils", "extdata", "pagexml-example.xml")
x        <- read_pagexml(path)
x

img      <- system.file(package = "madoc.utils", "extdata", "pagexml-example.jpg")
img      <- image_read(img)
plt      <- image_draw_polygons(img, x$coords, col = "#FF000080", border = "blue", lwd = 2)
plt
}
