% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_rbind}
\alias{image_rbind}
\title{Stack images below one another}
\usage{
image_rbind(image)
}
\arguments{
\item{image}{either an object of class \code{magick-image} or a character vector of files}
}
\value{
an object of class \code{magick-image} where all images are put below one another
}
\description{
Stack images below one another
}
\examples{
library(magick) 
library(data.table)
path     <- system.file(package = "madoc.utils", "extdata", "alto-example.xml")
x        <- read_alto(path)
x$width  <- as.integer(x$WIDTH)
x$height <- as.integer(x$HEIGHT)
x$x_left <- as.integer(x$HPOS)
x$y_top  <- as.integer(x$VPOS)

img      <- system.file(package = "madoc.utils", "extdata", "alto-example.jpg")
img      <- image_read(img)
areas    <- image_crop_textlines(img, x, color = "red")
areas$overview
areas$areas

all      <- image_rbind(areas$areas)
all
all      <- do.call(c, areas$areas)
all
}
