% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_canvas_model}
\alias{madoc_canvas_model}
\title{Retrieve canvas model annotations from Madoc}
\usage{
madoc_canvas_model(site, id)
}
\arguments{
\item{site}{character string with the site}

\item{id}{an id of a canvas}
}
\value{
in case 
\itemize{
\item{id is of length 1: }{a list with elements canvas_id and annotations where annotations is a data.frame with columns document_id, document_type, document_label, model_id, status, authors, id, type, value, label, id_revision, id_revises, selector_state, selector_type, selector_id. }
\item{id is vector of length > 1: }{a data.frame with columns canvas_id, document_id, document_type, document_label, model_id, status, authors, id, type, value, label, id_revision, id_revises, selector_state, selector_type, selector_id. Containing only rows for canvasses which have annotations.}
}
}
\description{
Get a data.frame of annotations by using the Madoc API's.\cr
This function performs a GET request to /madoc/api/canvases/{canvas_id}/models
}
\examples{
projects     <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije")
manifests    <- madoc_collection(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                                 id = projects$collection_id)

ids          <- sample(manifests$manifest_id, size = 10)
canvasses    <- madoc_manifest(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                               id = ids)
annotations  <- madoc_canvas_model(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                                   id = ids)
}
