% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api-with-token.R
\name{madoc_delete_user}
\alias{madoc_delete_user}
\title{Delete a user from Madoc}
\usage{
madoc_delete_user(id)
}
\arguments{
\item{id}{id of the user}
}
\value{
the result of the DELETE call
}
\description{
Delete a user from Madoc
}
\examples{
\dontrun{
madoc_login("https://www.madoc.ugent.be/s/brugse-vrije", 
            email = "jan.wijffels@vub.be", password = Sys.getenv("MADOC_PWD"))
madoc_delete_user(id = -1)
}
}
