% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_manifest}
\alias{madoc_manifest}
\title{Retrieve canvasses from Madoc}
\usage{
madoc_manifest(site, id)
}
\arguments{
\item{site}{character string with the site}

\item{id}{a vector of id's of manifests}
}
\value{
a data.frame with columns manifest_id, manifest_label, manifest_thumbnail, canvas_id, canvas_thumbnail, is_published
}
\description{
Get a data.frame of all canvasses which are part of a manifest by using the Madoc API's.\cr
This function performs a GET request to /madoc/api/manifests
}
\examples{
projects  <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije")
manifests <- madoc_collection(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                              id = projects$collection_id)

canvasses <- madoc_manifest(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                            id = sample(manifests$manifest_id, size = 1))
ids       <- sample(manifests$manifest_id, size = 10)
canvasses <- madoc_manifest(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                            id = ids)
}
