% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_manifest_metadata}
\alias{madoc_manifest_metadata}
\title{Put manifest metadata in a data.frame}
\usage{
madoc_manifest_metadata(x)
}
\arguments{
\item{x}{a data.frame with the output of \code{\link{madoc_manifest}}}
}
\value{
a data.frame with columns manifest_id and the columns available in the metadata, lowercased
}
\description{
Convert manifest metadata to a data.frame
}
\examples{
site      <- "https://www.madoc.ugent.be/s/brugse-vrije"
projects  <- madoc_projects(site)
manifests <- madoc_collection(site = site, id = projects$collection_id)
metadata  <- madoc_manifest_metadata(manifests)
}
