% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_projects}
\alias{madoc_projects}
\title{Retrieve projects from Madoc}
\usage{
madoc_projects(site)
}
\arguments{
\item{site}{character string with the site}
}
\value{
a data.frame with columns project_id, collection_id, slug, label and summary
}
\description{
Get a data.frame of all projects on a Madoc site by using the Madoc API's.\cr
This function performs a GET request to /madoc/api/projects.
}
\note{
TODO: perform pagination
}
\examples{
projects  <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije")
projects  <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije-gebruikerstest")
}
