% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api-with-token.R
\name{madoc_taskdetails}
\alias{madoc_taskdetails}
\title{Get task details}
\usage{
madoc_taskdetails(site, id)
}
\arguments{
\item{site}{character string with the main site (not project-specific)}

\item{id}{character string with the id of the task}
}
\value{
a data.frame with the Madoc task details
}
\description{
Get details on tasks. This requires logging in using \code{\link{madoc_login}}
}
\examples{
madoc_login("https://www.madoc.ugent.be/s/brugse-vrije", 
            email = "jan.wijffels@vub.be", password = Sys.getenv("MADOC_PWD"))
projects  <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije")
projects  <- subset(projects, slug == "brugse-vrije-gebruikerstest")
manifests <- madoc_collection(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                              id = projects$collection_id, tidy_metadata = TRUE)
canvasses <- madoc_manifest(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                            id = manifests$manifest_id)
ids       <- canvasses$canvas_id                   
ids       <- sample(canvasses$canvas_id, size = 5)
x         <- madoc_tasks(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                         project = "brugse-vrije-gebruikerstest", 
                         id = canvasses$canvas_id, type = "canvas")
x         <- subset(x, !is.na(id))
tk        <- madoc_taskdetails("https://www.madoc.ugent.be", x$id)
}
