% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api-with-token.R
\name{madoc_tasks}
\alias{madoc_tasks}
\title{TODO}
\usage{
madoc_tasks(site, project, id, type = c("manifest", "canvas"))
}
\arguments{
\item{site}{TODO}

\item{project}{TODO}

\item{id}{TODO}

\item{type}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
site    <- "https://www.madoc.ugent.be/s/brugse-vrije"
project <- "brugse-vrije-gebruikerstest"
id      <- 407
x       <- madoc_tasks(site = site, project = project, id = id, type = "manifest")
/s/:slug/madoc/api/projects/:projectSlug/manifest-tasks/:manifestId'
/s/:slug/madoc/api/projects/:projectSlug/canvas-tasks/:canvasId'
}
