% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alto.R
\name{read_alto}
\alias{read_alto}
\title{Read Alto-XML file}
\usage{
read_alto(x, type = c("transkribus"), ...)
}
\arguments{
\item{x}{path to the xml file}

\item{type}{character string with the type of xml structure. Defaults to 'transkribus' and is currently the only option.}

\item{...}{further arguments currently not used}
}
\value{
a data.frame
}
\description{
Import Alto-XML file as a data.frame. Examples where this can be used
is for importing alto-xml files from Transkribus
}
\examples{
f <- system.file(package = "madoc.utils", "extdata", "alto-example.xml")
x <- read_alto(f)
f <- system.file(package = "madoc.utils", "extdata", "altoxml-example.xml")
x <- read_alto(f)
}
