% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{image_crop_textpolygons}
\alias{image_crop_textpolygons}
\title{Crop out areas with text polygons from an image}
\usage{
image_crop_textpolygons(
  image,
  geometries,
  color = "royalblue",
  border = "10x10",
  overview = TRUE,
  max_width = +Inf,
  trace = FALSE
)
}
\arguments{
\item{image}{\code{image} either an object of class \code{opencv-image} or a path to an image file on disk}

\item{geometries}{a data.frame with columns width, height, x_left, y_top indicating the areas to extract from the image}

\item{color}{color to use for adding a border in the overview image. Defaults to 'royalblue'.}

\item{border}{border pixels to using in the overview image. Defaults to 10x10 pixel borders.}

\item{overview}{logical indicating to add the overview image of all area's below each other. Defaults to TRUE.}

\item{max_width}{maximum width of the overview image. Defaults to +Inf}

\item{trace}{logical indicating to trace progress}
}
\value{
a list with elements areas and overview where \code{overview} is a \code{magick-image} with stacked image lines
and \code{areas} is a list of \code{magick-image}'s, one for each text line \cr
In case overview is set to \code{FALSE} the return value is only the list of stacked image lines.
}
\description{
Crop out areas with text polygons from an image
}
\examples{
library(opencv) 
library(data.table)
path     <- system.file(package = "madoc.utils", "extdata", "pagexml-example.xml")
x        <- read_pagexml(path)

img      <- system.file(package = "madoc.utils", "extdata", "pagexml-example.jpg")
img      <- ocv_read(img)
area     <- ocv_polygon(img, pts = x$coords[[5]])
areas    <- image_crop_textpolygons(img, x, color = "red")
areas$overview
areas$areas
}
