% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_canvas_image}
\alias{madoc_canvas_image}
\title{Retrieve canvas image details from Madoc}
\usage{
madoc_canvas_image(site, id)
}
\arguments{
\item{site}{character string with the site}

\item{id}{a vector of id's of canvasses}
}
\value{
a data.frame with columns canvas_id, height, width, image_url
}
\description{
Get a data.frame of canvas information (image height/width + image_url) by using the Madoc API's.\cr
This function performs a GET request to /madoc/api/canvases
}
\examples{
projects  <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije")
manifests <- madoc_collection(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                              id = projects$collection_id)

ids               <- sample(manifests$manifest_id, size = 10)
canvasses         <- madoc_manifest(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                                    id = ids)
canvasses_urls    <- madoc_canvas_image("https://www.madoc.ugent.be/s/brugse-vrije", 
                                        id = canvasses$canvas_id)
canvasses         <- merge(canvasses, canvasses_urls, by = "canvas_id")

library(magick)
image_read(unlist(canvasses$image_url[[1]]))
}
