% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_collection}
\alias{madoc_collection}
\title{Retrieve a collection from Madoc}
\usage{
madoc_collection(site, id)
}
\arguments{
\item{site}{character string with the site}

\item{id}{the id of the collection}
}
\value{
a data.frame with columns collection_id, manifest_id, type, label, canvasCount, thumbnail and metadata
}
\description{
Get a data.frame of all manifests which are part of a collection by using the Madoc API's.\cr
This function performs a GET request to /madoc/api/collections and handles pagination
}
\examples{
site      <- "https://www.madoc.ugent.be/s/brugse-vrije"
projects  <- madoc_projects(site)
manifests <- madoc_collection(site = site, id = projects$collection_id)
}
