% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api.R
\name{madoc_tasks}
\alias{madoc_tasks}
\title{Get Madoc tasks}
\usage{
madoc_tasks(site, project, id, type = c("manifest", "canvas"))
}
\arguments{
\item{site}{character string with the site}

\item{project}{character string with the project name of the site}

\item{id}{a vector of manifest or task identifiers}

\item{type}{character string with the type of task, either 'manifest' or 'canvas'}
}
\value{
a data.frame with task information
}
\description{
Retrieve tasks from Madoc, namely manifest and canvas tass
}
\examples{
projects  <- madoc_projects("https://www.madoc.ugent.be/s/brugse-vrije")
projects  <- subset(projects, slug == "brugse-vrije-gebruikerstest")
manifests <- madoc_collection(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                              id = projects$collection_id, tidy_metadata = TRUE)
canvasses <- madoc_manifest(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                            id = manifests$manifest_id)
ids       <- sample(manifests$manifest_id, size = 5)
x         <- madoc_tasks(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                         project = "brugse-vrije-gebruikerstest", 
                         id = ids, type = "manifest")
ids       <- sample(canvasses$canvas_id, size = 5)
x         <- madoc_tasks(site = "https://www.madoc.ugent.be/s/brugse-vrije", 
                         project = "brugse-vrije-gebruikerstest", 
                         id = canvasses$canvas_id, type = "canvas")
}
