% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madoc-api-with-token.R
\name{madoc_users}
\alias{madoc_users}
\title{Get the list of Madoc users}
\usage{
madoc_users()
}
\value{
a data.frame with the Madoc users with columns id, name, role, site_role, email
}
\description{
Get the list of Madoc users. This requires logging in using \code{\link{madoc_login}}
}
\examples{
madoc_login("https://www.madoc.ugent.be/s/brugse-vrije", 
            email = "jan.wijffels@vub.be", password = Sys.getenv("MADOC_PWD"))
x <- madoc_users()
}
