% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transkribus.R
\name{Transkribus}
\alias{Transkribus}
\title{Transkribus API}
\description{
Connect to Transkribus, inspect collections, documents, pages, perform handwritten text recognition
}
\examples{
library(madoc.utils)
api <- Transkribus$new(user     = "jan.wijffels@vub.ac.be", 
                       password = Sys.getenv("TRANSKRIBUS_PWD"))

## Get pages of a collection
collections   <- api$list_collections()
collections
id_collection <- sample(collections$colId, size = 1)
documents     <- api$list_collection(collection = id_collection)
documents
id_document   <- sample(documents$docId, size = 1)
pages         <- api$list_document(collection = id_collection, document = id_document)
pages

## Create a collection, upload some images to the collection, delete it again
id  <- api$create_collection(label = "example-collection")
img <- c(system.file(package = "madoc.utils", "extdata", "alto-example.jpg"),
         system.file(package = "madoc.utils", "extdata", "example.png"))
api$upload(data = img, collection = id, document = paste("Upload", Sys.time()), author = "R-API")
api$list_collection(collection = id)           
api$delete_collection(collection = id)

## Look at relevant models and dictionaries
dicts    <- api$list_dictionaries()
grep(dicts, pattern = "Dutch", ignore.case = TRUE, value = TRUE)
models   <- api$list_models(collection = id_collection)
str(models)
dutch    <- grep(models$language, pattern = "Dutch", ignore.case = TRUE, value = TRUE)
dutch    <- subset(models, language \%in\% dutch)
dutch    <- c("Dutch Mountains (18th Century)", "IJsberg", "Dutch Notarial Model 18th Century")
dutch    <- subset(models, name \%in\% dutch)
dutch    <- subset(models, name \%in\% "Dutch Mountains (18th Century)" & provider == "CITlabPlus")
str(dutch)
id_model <- dutch$htrId

## Inspect jobs
jobs   <- api$list_job()
jobs


\dontrun{
id  <- api$create_collection(label = "test-collection")
img <- c(system.file(package = "madoc.utils", "extdata", "example.png"),
         system.file(package = "madoc.utils", "extdata", "alto-example.jpg"))
api$upload(data = img, collection = id, document = "Doc with 2 images", author = "R-API")

##
## This section shows how to transcribe using the API
##    >> note that this consumes Transkribus credits

##
## Inspect one image and transcribe it
##
##  - id_model      <- 21683  ## Dutch Mountains HTR+
id_collection <- id
id_document   <- docs$docId
docs          <- api$list_collection(collection = id_collection)
pages         <- api$list_document(collection = id_collection, document = id_document)
page          <- head(pages, n = 1)
id_job        <- api$transcribe(collection = id_collection, document = id_document, 
                                page = 1,
                                model = 21683, 
                                dictionary = "Combined_Dutch_Model_M1.dict")
x             <- read_pagexml(page$page_xml) 

##
## A random document from a collection
##
library(magick)
pages  <- api$list_document(collection = id_collection, document = id_document)
page   <- tail(pages, n = 1)
page
img    <- image_read(page$thumbUrl)
img    <- image_read(page$url)
image_resize(img, "x600")
id_job <- api$transcribe(collection = id_collection, document = id_document, page = page$pageNr, 
                         model = id_model, dictionary = "Combined_Dutch_Model_M1.dict")
api$list_job(job = id_job)   
 
## After the job has finished, we have a Page-XML file which we can read in
pages  <- api$list_document(collection = id_collection, document = id_document)  
page   <- tail(pages, n = 1)
img    <- image_read(page$url)
x      <- read_pagexml(page$page_xml)  
bl     <- image_draw_baselines(img, x = x$baseline, col = "darkgreen", lwd = 4)
image_resize(bl, "x900")
bl    <- image_crop_baselineareas(img, 
                                  x = setNames(x$baseline, x$id), 
                                  textregion = x$points, 
                                  extend = FALSE, overview = FALSE)
bl    <- image_rbind(bl, color = "red", geometry = "2x2")   
image_resize(bl, "x900")

api$delete_collection(collection = id)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{JSESSIONID}}{character string with the JSESSIONID to use once logged in}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Transkribus$new()}}
\item \href{#method-map_name_to_id}{\code{Transkribus$map_name_to_id()}}
\item \href{#method-list_collections}{\code{Transkribus$list_collections()}}
\item \href{#method-list_collection}{\code{Transkribus$list_collection()}}
\item \href{#method-create_collection}{\code{Transkribus$create_collection()}}
\item \href{#method-delete_collection}{\code{Transkribus$delete_collection()}}
\item \href{#method-list_document}{\code{Transkribus$list_document()}}
\item \href{#method-list_dictionaries}{\code{Transkribus$list_dictionaries()}}
\item \href{#method-list_models}{\code{Transkribus$list_models()}}
\item \href{#method-list_job}{\code{Transkribus$list_job()}}
\item \href{#method-transcribe}{\code{Transkribus$transcribe()}}
\item \href{#method-upload}{\code{Transkribus$upload()}}
\item \href{#method-clone}{\code{Transkribus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Log in with your Transkribus user and password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$new(
  url = "https://transkribus.eu/TrpServer/rest/auth/login",
  user,
  password
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{user}}{character string with your Transkribus user in order to connect}

\item{\code{password}}{character string with your Transkribus password in order to connect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_name_to_id"></a>}}
\if{latex}{\out{\hypertarget{method-map_name_to_id}{}}}
\subsection{Method \code{map_name_to_id()}}{
Map labels to identifiers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$map_name_to_id(
  name,
  type = c("collection", "document"),
  collection
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character string with the label to map to the identifier}

\item{\code{type}}{type of mapping: either one of 'collection' of 'document'}

\item{\code{collection}}{id of the collection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_collections"></a>}}
\if{latex}{\out{\hypertarget{method-list_collections}{}}}
\subsection{Method \code{list_collections()}}{
List all collections you have access to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$list_collections(
  url = "https://transkribus.eu/TrpServer/rest/collections/list"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_collection"></a>}}
\if{latex}{\out{\hypertarget{method-list_collection}{}}}
\subsection{Method \code{list_collection()}}{
List the content (the documents) of a collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$list_collection(
  url = "https://transkribus.eu/TrpServer/rest/collections/\%s/list",
  collection
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{collection}}{id of the collection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_collection"></a>}}
\if{latex}{\out{\hypertarget{method-create_collection}{}}}
\subsection{Method \code{create_collection()}}{
Create a collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$create_collection(
  url = "https://transkribus.eu/TrpServer/rest/collections/createCollection",
  label
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{label}}{character string with the name of the collection to create}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_collection"></a>}}
\if{latex}{\out{\hypertarget{method-delete_collection}{}}}
\subsection{Method \code{delete_collection()}}{
Delete a collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$delete_collection(
  url = "https://transkribus.eu/TrpServer/rest/collections/{collection}",
  collection
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{collection}}{id of the collection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_document"></a>}}
\if{latex}{\out{\hypertarget{method-list_document}{}}}
\subsection{Method \code{list_document()}}{
List the content (the pages) of a document
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$list_document(
  url = "https://transkribus.eu/TrpServer/rest/collections/\%s/\%s/fulldoc",
  collection,
  document,
  type = c("pages", "raw")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{collection}}{id of the collection}

\item{\code{document}}{id of the document}

\item{\code{type}}{character string with the type of extraction, either 'pages' or 'raw'. Defaults to 'pages'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_dictionaries"></a>}}
\if{latex}{\out{\hypertarget{method-list_dictionaries}{}}}
\subsection{Method \code{list_dictionaries()}}{
Retrieve the set of dictionaries containing possible letters as output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$list_dictionaries(
  url = "https://transkribus.eu/TrpServer/rest/recognition/dicts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_models"></a>}}
\if{latex}{\out{\hypertarget{method-list_models}{}}}
\subsection{Method \code{list_models()}}{
Retrieve all HTR/OCR models you have access to within a collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$list_models(
  url = "https://transkribus.eu/TrpServer/rest/recognition/\%s/list",
  collection
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{collection}}{id of the collection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_job"></a>}}
\if{latex}{\out{\hypertarget{method-list_job}{}}}
\subsection{Method \code{list_job()}}{
List all jobs or get the information of one specific job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$list_job(
  url = "https://transkribus.eu/TrpServer/rest/jobs/list",
  job
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{job}}{id of the job}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transcribe"></a>}}
\if{latex}{\out{\hypertarget{method-transcribe}{}}}
\subsection{Method \code{transcribe()}}{
Transcribe a set of pages with a model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$transcribe(
 
    url = "https://transkribus.eu/TrpServer/rest/recognition/{collection}/{model}/htrCITlab?id={document}&pages={page}&dict={dictionary}",
  collection,
  document,
  page,
  model,
  dictionary
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{collection}}{id of the collection}

\item{\code{document}}{id of the document}

\item{\code{page}}{id of the page to transcribe}

\item{\code{model}}{id of the Transkribus model to use}

\item{\code{dictionary}}{character string with the dictionary (set of letters) to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload"></a>}}
\if{latex}{\out{\hypertarget{method-upload}{}}}
\subsection{Method \code{upload()}}{
Upload a set of images in a collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$upload(
  url = c("https://transkribus.eu/TrpServer/rest/uploads?collId={collection}",
    "https://transkribus.eu/TrpServer/rest/uploads/{uploadId}"),
  collection,
  data,
  document,
  author = "R-API",
  trace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{character string with the url to use in the call to the Transkribus API}

\item{\code{collection}}{id of the collection}

\item{\code{data}}{a character vector with the full path(s) to the image files on disk}

\item{\code{document}}{the title of the document}

\item{\code{author}}{the author of the document}

\item{\code{trace}}{logical indicating to show progress}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transkribus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
